requirements.txt
python-dotenv==1.0.1
Pillow==10.3.0
numpy==1.26.4
predictionguard==2.1.0
opencv-python==4.10.0.82
webvtt-py==0.5.1
moviepy==1.0.3
pytube==15.0.0
pytubefix==6.5.2
youtube-transcript-api==0.6.2
langchain==0.2.4
langchain-community==0.2.4
langchain-core==0.2.6
langchain-text-splitters==0.2.1
pandas==2.2.2
lancedb==0.8.2
scikit-learn==1.5.1
umap-learn==0.5.6
matplotlib==3.9.1
seaborn==0.13.2
datasets==2.20.0
gradio==4.44.0
openai-whisper
#sudo apt-get install libglib2.0-0 libsm6 libxrender1 libxext6
